refactor: move 'costs' utils into 'swap' feature
